requirements.txt
alembic==1.4.2
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.14.31
botocore==1.17.31
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
dnspython==1.16.0
docutils==0.15.2
dominate==2.5.1
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gensim==3.8.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.0
murmurhash==1.0.2
nltk==3.5
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-futures==1.0.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.5.2
sentence-transformers==0.3.2
sentencepiece==0.1.91
six==1.15.0
smart-open==2.1.0
SQLAlchemy==1.3.18
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
torch==1.6.0
tqdm==4.48.0
transformers==3.0.2
urllib3==1.25.10
visitor==0.1.3
wasabi==0.7.1
Werkzeug==1.0.1
WTForms==2.3.1
